+++ src/main/java/org/bankofspring/dao/jdbc/CustomerDAOIJDBCmpl.java
@@ -1,5 +1,6 @@
-package org.bankofspring.dao;
+package org.bankofspring.dao.jdbc;
 
+import org.bankofspring.dao.CustomerDAO;
 import org.bankofspring.dao.jdbc.rowmapper.CustomerRowMapper;
 import org.bankofspring.model.Customer;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -7,7 +8,7 @@ import org.springframework.jdbc.core.simple.SimpleJdbcTemplate;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public class CustomerDAOImpl implements CustomerDAO {
+public class CustomerDAOIJDBCmpl implements CustomerDAO {
 
 	@Autowired
 	private SimpleJdbcTemplate jdbc;
