+++ src/test/java/org/bankofspring/dao/jdbc/CustomerDAOJDBCTest.java
@@ -1,7 +1,8 @@
-package org.bankofspring.dao;
+package org.bankofspring.dao.jdbc;
 
 import static org.junit.Assert.*;
 
+import org.bankofspring.dao.CustomerDAO;
 import org.bankofspring.model.Customer;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -14,7 +15,7 @@ import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 @RunWith(SpringJUnit4ClassRunner.class)
 @ContextConfiguration({"classpath:BankOfSpring-ds-test.xml", "classpath:BankOfSpring.xml"})
 @DirtiesContext(classMode = ClassMode.AFTER_EACH_TEST_METHOD)
-public class CustomerDAOTest {
+public class CustomerDAOJDBCTest {
 
 	@Autowired
 	private CustomerDAO customerDAO;
